# Fix AST inconsistency between SASS/SCSS caused by indentation