--- a/raffia/tests/ast/sass/indent-with-spaces.sass
+++ b/raffia/tests/ast/sass/indent-with-spaces.sass
@@ -4,3 +4,14 @@ a
 
 x
   y: z
+
+e
+  f
+    g
+      h
+        j
+          a: b
+    k
+l
+m
+n