--- a/raffia/tests/ast/sass/indent-with-spaces.sass
+++ b/raffia/tests/ast/sass/indent-with-spaces.sass
@@ -1,17 +1,10 @@
 a
- b: c
-    d: e
-
-x
-  y: z
-
-e
-  f
-    g
-      h
-        j
-          a: b
-    k
-l
-m
-n
+    color: red
+    b
+        c
+            d
+                color: red
+        color: red
+    e
+        color: red
+    foo