--- a/raffia/tests/ast/sass/indent-with-tabs.sass
+++ b/raffia/tests/ast/sass/indent-with-tabs.sass
@@ -4,3 +4,11 @@ a
 
 x
 		y: z
+
+a
+	b
+			c
+				d
+	e
+		f
+g