--- a/raffia/tests/ast/scss/rule-nesting.scss
+++ b/raffia/tests/ast/scss/rule-nesting.scss
@@ -1,0 +1,15 @@
+a {
+    color: red;
+    b {
+        c {
+            d {
+                color: red;
+            }
+        }
+        color: red;
+    }
+    e {
+        color: red;
+    }
+    foo {}
+}