Fixes #7 by

- Don't consume the source when `Dedent` token should be returned.
- Keep `Dedent` tokens in the state, because they don't have anything to actually consume
- Keep indentation size per level to know how much `Dedent`s should be generated.
    - make "inconsistent indentation" unrecoverable error

At 522d0ca1e6cd884d46e2047c747b4f025a088aa7 new ast's are in snapshots. Also added scss test to compare ASTs - seems fine, plz check :3

Not sure about 24e70c358d9f919ad54f2562ca6e37a3e992ba85 though. Looks premature..?