@g-plane 2024-11-05
We are cloning state in many cases. Does changing the definition of `TokenizerState` affect performance?