@yiffyrusdev 2024-11-05
It would be as much slower, as two heap allocations of (n) + 2 `Vec` triplets are slow, AFAIK. I'd say that N is (2+16+3)*indentation bytes (2 for u16 for level, 16 is Span, 3 is the "worst" case for the Token I've found)

We *could* use the `SmallVec` which seems to be faster for the case, but the potential indentation depth must be pre-estimated, which is a magic number, similar to the capacity of "8" 24e70c358d9f919ad54f2562ca6e37a3e992ba85

<details>
<summary>I did 3 tests before and one after the changes:</summary>

Here are two last of them, those come from criterion for parse_selector/Raffia
```
<before>

Lower                    Bound	       Estimate	       Upper Bound
Value:	                  1.28µs	1.28µs	        1.28µs
Change in Value:	-0.6323%	-0.0569%	+0.4337%
Change within noise threshold.

<after>
	                Lower Bound	Estimate	Upper Bound
Value:	                 1.26µs	          1.26µs	1.26µs
Change in Value:	-1.9907%	-1.8563%	-1.7194%
No change in performance detected.
```
i9-13980HX
</details>