@g-plane 2024-11-05
You can run the "self" benchmark which only runs for Raffia itself and supports CSS, SCSS, Sass and Less. Other benchmarks are only for CSS.