@yiffyrusdev 2024-11-05
> Other benchmarks are only for CSS and they will run on other parsers

Oh, my bad, I completely missed that. Then It's ~(+17%) for bootstrap5/formcheck.sass. Definitely must be better.

Pardon me for the PR-noise then, I'll convert this to draft for now, if it's OK.